+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-# [9.1.1] - 2025-04-30
-- Allow setting `HttpClient` on `HttpWrapper`
+# [9.2.0] - 2025-04-30
+- Added support for setting additional request headers in `HttpConfig`
+- Allow setting `HttpClient` and `HttpConfig` on `HttpWrapper`
 - Bumped Jackson version to 2.19.0
 
 # [9.1.0] - 2025-04-22